the new design is finisch & next phase is bugfixes...
